Modularize AdminPanelStackNavigator
